refactor(copyright): split author heuristics modules
