+++ src/copyright/detector/author_heuristics/mod.rs
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: Provenant contributors
+// SPDX-License-Identifier: Apache-2.0
+
+#[cfg(test)]
+#[path = "../author_heuristics_test.rs"]
+mod tests;
+
+mod cleanup;
+mod extraction;
+
+pub(super) use cleanup::*;
+pub(super) use extraction::*;
